[REF] Menu: Split the Menu from its Popover
Extract the Menu logic in a given component to use it outside of a
popover.
