+++ src/components/bottom_bar/bottom_bar.xml
@@ -87,7 +87,7 @@
         />
       </t>
 
-      <Menu
+      <MenuPopover
         t-if="menuState.isOpen"
         anchorRect="menuState.anchorRect"
         menuItems="menuState.menuItems"
