+++ src/components/figures/chart/chart_dashboard_menu/chart_dashboard_menu.xml
@@ -16,7 +16,7 @@
           t-on-click="openContextMenu"
         />
       </div>
-      <Menu
+      <MenuPopover
         t-if="menuState.isOpen"
         anchorRect="menuState.anchorRect"
         menuItems="menuState.menuItems"
