+++ src/components/figures/figure/figure.xml
@@ -26,7 +26,7 @@
             t-on-contextmenu.prevent.stop="showMenu">
             <t t-call="o-spreadsheet-Icon.LIST"/>
           </div>
-          <Menu
+          <MenuPopover
             t-if="menuState.isOpen"
             anchorRect="menuState.anchorRect"
             menuItems="menuState.menuItems"
