+++ src/components/grid/grid.xml
@@ -53,7 +53,7 @@
         </t>
       </t>
       <Selection t-if="displaySelectionHandler"/>
-      <Menu
+      <MenuPopover
         t-if="menuState.isOpen"
         menuItems="menuState.menuItems"
         anchorRect="menuState.anchorRect"
