+++ src/components/menu/menu.ts
@@ -1,11 +1,4 @@
-import {
-  Component,
-  onWillUnmount,
-  onWillUpdateProps,
-  useExternalListener,
-  useRef,
-  useState,
-} from "@odoo/owl";
+import { Component, onWillUnmount } from "@odoo/owl";
 import { Action } from "../../actions/action";
 import {
   BUTTON_ACTIVE_BG,
@@ -15,20 +8,9 @@ import {
   MENU_ITEM_HEIGHT,
   MENU_ITEM_PADDING_HORIZONTAL,
   MENU_ITEM_PADDING_VERTICAL,
-  MENU_VERTICAL_PADDING,
-  MENU_WIDTH,
 } from "../../constants";
-import { MenuMouseEvent, Pixel, Rect, SpreadsheetChildEnv, UID } from "../../types";
-import { PopoverPropsPosition } from "../../types/cell_popovers";
+import { DOMCoordinates, Pixel, Rect, SpreadsheetChildEnv } from "../../types";
 import { css, cssPropertiesToCss } from "../helpers/css";
-import {
-  getOpenedMenus,
-  getRefBoundingRect,
-  isChildEvent,
-  isMiddleClickOrCtrlClick,
-} from "../helpers/dom_helpers";
-import { useTimeOut } from "../helpers/time_hooks";
-import { Popover, PopoverProps } from "../popover/popover";
 
 //------------------------------------------------------------------------------
 // Context Menu Component
@@ -37,8 +19,6 @@ import { Popover, PopoverProps } from "../popover/popover";
 css/* scss */ `
   .o-menu {
     background-color: white;
-    padding: ${MENU_VERTICAL_PADDING}px 0px;
-    width: ${MENU_WIDTH}px;
     user-select: none;
 
     .o-menu-item {
@@ -83,18 +63,16 @@ css/* scss */ `
 
 type MenuItemOrSeparator = Action | "separator";
 
-const TIMEOUT_DELAY = 250;
-
-interface Props {
-  anchorRect: Rect;
-  popoverPositioning: PopoverPropsPosition;
+export interface MenuProps {
+  position: DOMCoordinates;
   menuItems: Action[];
-  depth: number;
-  maxHeight?: Pixel;
   onClose: () => void;
-  onMenuClicked?: (ev: CustomEvent) => void;
-  menuId?: UID;
-  onMouseOver?: () => void;
+  onScroll?: (ev: CustomEvent) => void;
+  onClickMenu?: (menu: Action, ev: CustomEvent) => void;
+  onMouseEnter?: (menu: Action, ev: PointerEvent) => void;
+  onMouseOver?: (menu: Action, ev: PointerEvent) => void;
+  onMouseLeave?: (menu: Action, ev: PointerEvent) => void;
+  isActive?: (menu: Action) => boolean;
   width?: number;
 }
 
@@ -107,46 +85,26 @@ export interface MenuState {
   isHoveringChild?: boolean;
 }
 
-export class Menu extends Component<Props, SpreadsheetChildEnv> {
+export class Menu extends Component<MenuProps, SpreadsheetChildEnv> {
   static template = "o-spreadsheet-Menu";
   static props = {
-    anchorRect: Object,
-    popoverPositioning: { type: String, optional: true },
     menuItems: Array,
-    depth: { type: Number, optional: true },
-    maxHeight: { type: Number, optional: true },
     onClose: Function,
-    onMenuClicked: { type: Function, optional: true },
-    menuId: { type: String, optional: true },
+    onClickMenu: { type: Function, optional: true },
+    onMouseEnter: { type: Function, optional: true },
     onMouseOver: { type: Function, optional: true },
+    onMouseLeave: { type: Function, optional: true },
     width: { type: Number, optional: true },
+    isActive: { type: Function, optional: true },
+    onScroll: { type: Function, optional: true },
   };
 
-  static components = { Menu, Popover };
-  static defaultProps = {
-    depth: 1,
-    popoverPositioning: "top-right",
-  };
-  private subMenu: MenuState = useState({
-    isOpen: false,
-    anchorRect: null,
-    scrollOffset: 0,
-    menuItems: [],
-    isHoveringChild: false,
-  });
-  private menuRef = useRef("menu");
+  static components = {};
+  static defaultProps = {};
+
   private hoveredMenu: Action | undefined = undefined;
 
-  private openingTimeOut = useTimeOut();
-
   setup() {
-    useExternalListener(window, "click", this.onExternalClick, { capture: true });
-    useExternalListener(window, "contextmenu", this.onExternalClick, { capture: true });
-    onWillUpdateProps((nextProps: Props) => {
-      if (nextProps.menuItems !== this.props.menuItems) {
-        this.closeSubMenu();
-      }
-    });
     onWillUnmount(() => {
       this.hoveredMenu?.onStopHover?.(this.env);
     });
@@ -176,28 +134,6 @@ export class Menu extends Component<Props, SpreadsheetChildEnv> {
     return menuItemsAndSeparators;
   }
 
-  get subMenuAnchorRect(): Rect {
-    const anchorRect = Object.assign({}, this.subMenu.anchorRect);
-    anchorRect.y -= this.subMenu.scrollOffset || 0;
-    return anchorRect;
-  }
-
-  get popoverProps(): PopoverProps {
-    const isRoot = this.props.depth === 1;
-    return {
-      anchorRect: {
-        x: this.props.anchorRect.x,
-        y: this.props.anchorRect.y,
-        width: isRoot ? this.props.anchorRect.width : this.props.width || MENU_WIDTH,
-        height: isRoot ? this.props.anchorRect.height : MENU_ITEM_HEIGHT,
-      },
-      positioning: this.props.popoverPositioning,
-      verticalOffset: isRoot ? 0 : MENU_VERTICAL_PADDING,
-      onPopoverHidden: () => this.closeSubMenu(),
-      onPopoverMoved: () => this.closeSubMenu(),
-    };
-  }
-
   get childrenHaveIcon(): boolean {
     return this.props.menuItems.some((menuItem) => !!this.getIconName(menuItem));
   }
@@ -221,27 +157,6 @@ export class Menu extends Component<Props, SpreadsheetChildEnv> {
     return cssPropertiesToCss({ color: menu.iconColor });
   }
 
-  async activateMenu(menu: Action, isMiddleClick?: boolean) {
-    const result = await menu.execute?.(this.env, isMiddleClick);
-    this.close();
-    this.props.onMenuClicked?.({ detail: result } as CustomEvent);
-  }
-
-  private close() {
-    this.closeSubMenu();
-    this.props.onClose();
-  }
-
-  private onExternalClick(ev: MenuMouseEvent) {
-    // Don't close a root menu when clicked to open the submenus.
-    const el = this.menuRef.el;
-    if (el && getOpenedMenus().some((el) => isChildEvent(el, ev))) {
-      return;
-    }
-    ev.closedMenuId = this.props.menuId;
-    this.close();
-  }
-
   getName(menu: Action) {
     return menu.name(this.env);
   }
@@ -257,96 +172,19 @@ export class Menu extends Component<Props, SpreadsheetChildEnv> {
     return false;
   }
 
-  isActive(menuItem: Action): boolean {
-    return (this.subMenu?.isHoveringChild || false) && this.isParentMenu(this.subMenu, menuItem);
-  }
-
-  onScroll(ev) {
-    this.subMenu.scrollOffset = ev.target.scrollTop;
-  }
-
-  /**
-   * If the given menu is not disabled, open it's submenu at the
-   * correct position according to available surrounding space.
-   */
-  private openSubMenu(menu: Action, parentMenuEl: HTMLElement) {
-    if (!parentMenuEl) {
-      return;
-    }
-    const y = parentMenuEl.getBoundingClientRect().top;
-
-    this.subMenu.anchorRect = {
-      x: getRefBoundingRect(this.menuRef).x,
-      y: y - (this.subMenu.scrollOffset || 0),
-      width: this.props.width || MENU_WIDTH,
-      height: MENU_ITEM_HEIGHT,
-    };
-    this.subMenu.menuItems = menu.children(this.env);
-    this.subMenu.isOpen = true;
-    this.subMenu.parentMenu = menu;
-  }
-
-  isParentMenu(subMenu: MenuState, menuItem: Action) {
-    return subMenu.parentMenu?.id === menuItem.id;
-  }
-
-  private closeSubMenu() {
-    if (this.subMenu.isHoveringChild) {
-      return;
-    }
-    this.subMenu.isOpen = false;
-    this.subMenu.parentMenu = undefined;
-  }
-
-  onClickMenu(menu: Action, ev: MouseEvent) {
-    if (!this.isEnabled(menu)) {
-      return;
-    }
-
-    if (this.isRoot(menu)) {
-      this.openSubMenu(menu, ev.currentTarget as HTMLElement);
-    } else {
-      this.activateMenu(menu, isMiddleClickOrCtrlClick(ev));
-    }
-  }
-
-  onMouseOver(menu: Action, ev: MouseEvent) {
-    if (this.isEnabled(menu)) {
-      if (this.isParentMenu(this.subMenu, menu)) {
-        this.openingTimeOut.clear();
-        return;
-      }
-      const currentTarget = ev.currentTarget as HTMLElement;
-      if (this.isRoot(menu)) {
-        this.openingTimeOut.schedule(() => {
-          this.openSubMenu(menu, currentTarget);
-        }, TIMEOUT_DELAY);
-      }
-    }
-  }
-
-  onMouseOverMainMenu() {
-    this.props.onMouseOver?.();
-    this.subMenu.isHoveringChild = false;
-  }
-
-  onMouseOverChildMenu() {
-    this.subMenu.isHoveringChild = true;
-    this.openingTimeOut.clear();
-  }
-
-  onMouseEnter(menu: Action, ev: MouseEvent) {
-    this.hoveredMenu = menu;
-    menu.onStartHover?.(this.env);
-  }
-
-  onMouseLeave(menu: Action) {
-    this.openingTimeOut.schedule(this.closeSubMenu.bind(this), TIMEOUT_DELAY);
-    this.hoveredMenu = undefined;
-    menu.onStopHover?.(this.env);
-  }
-
   get menuStyle() {
     return this.props.width ? cssPropertiesToCss({ width: this.props.width + "px" }) : "";
+  }
+
+  onMouseEnter(menu: Action, ev: PointerEvent) {
+    this.hoveredMenu = menu;
+    menu.onStartHover?.(this.env);
+    this.props.onMouseEnter?.(menu, ev);
+  }
+
+  onMouseLeave(menu: Action, ev: PointerEvent) {
+    this.hoveredMenu = undefined;
+    menu.onStopHover?.(this.env);
+    this.props.onMouseLeave?.(menu, ev);
   }
 }
