+++ src/components/side_panel/components/cog_wheel_menu/cog_wheel_menu.ts
@@ -3,7 +3,7 @@ import { ActionSpec, createActions } from "../../../../actions/action";
 import { MenuMouseEvent } from "../../../../types";
 import { SpreadsheetChildEnv } from "../../../../types/env";
 import { getBoundingRectAsPOJO } from "../../../helpers/dom_helpers";
-import { Menu, MenuState } from "../../../menu/menu";
+import { MenuPopover, MenuState } from "../../../menu_popover/menu_popover";
 
 interface Props {
   items: ActionSpec[];
@@ -11,7 +11,7 @@ interface Props {
 
 export class CogWheelMenu extends Component<Props, SpreadsheetChildEnv> {
   static template = "o-spreadsheet-CogWheelMenu";
-  static components = { Menu };
+  static components = { MenuPopover };
   static props = {
     items: Array,
   };
