+++ src/components/side_panel/components/cog_wheel_menu/cog_wheel_menu.xml
@@ -5,7 +5,7 @@
       t-on-click="toggleMenu"
       t-ref="button"
     />
-    <Menu
+    <MenuPopover
       t-if="menuState.isOpen"
       menuId="menuId"
       anchorRect="menuState.anchorRect"
