+++ src/components/side_panel/select_menu/select_menu.xml
@@ -7,7 +7,7 @@
       t-on-click="onClick">
       <option selected="true" t-esc="props.selectedValue"/>
     </select>
-    <Menu
+    <MenuPopover
       t-if="state.isMenuOpen"
       menuItems="props.menuItems"
       anchorRect="menuAnchorRect"
