+++ src/components/top_bar/number_formats_tool/number_formats_tool.ts
@@ -5,7 +5,7 @@ import { Rect, SpreadsheetChildEnv } from "../../../types";
 import { ActionButton } from "../../action_button/action_button";
 import { getBoundingRectAsPOJO } from "../../helpers/dom_helpers";
 import { ToolBarDropdownStore, useToolBarDropdownStore } from "../../helpers/top_bar_tool_hook";
-import { Menu } from "../../menu/menu";
+import { MenuPopover } from "../../menu_popover/menu_popover";
 
 interface Props {
   class: string;
@@ -18,7 +18,7 @@ interface State {
 
 export class NumberFormatsTool extends Component<Props, SpreadsheetChildEnv> {
   static template = "o-spreadsheet-NumberFormatsTool";
-  static components = { Menu, ActionButton };
+  static components = { MenuPopover, ActionButton };
   static props = { class: String };
   formatNumberMenuItemSpec = formatNumberMenuItemSpec;
   topBarToolStore!: ToolBarDropdownStore;
