+++ src/components/top_bar/number_formats_tool/number_formats_tool.xml
@@ -6,7 +6,7 @@
       onClick.bind="toggleMenu"
       class="props.class"
     />
-    <Menu
+    <MenuPopover
       t-if="isActive"
       anchorRect="state.anchorRect"
       menuItems="state.menuItems"
